requirements.txt
fastapi
google-generativeai
python-dotenv
uvicorn